Restructure AGFS demo claim generators
