+++ lib/demo_data/claim_generator/agfs/advocate.rb
@@ -1,0 +1,8 @@
+module DemoData
+  class ClaimGenerator
+    class AGFS
+      class Advocate < AGFS
+      end
+    end
+  end
+end
